refactor: extract reusable driver widgets
